Refinement of upb_bytesrc interface.
Added a upb_byteregion that tracks a region of
the input buffer; decoders use this instead of
using a upb_bytesrc directly.  upb_byteregion
is also used as the way of passing a string to
a upb_handlers callback.  This symmetry makes
decoders compose better; if you want to take
a parsed string and decode it as something else,
you can take the string directly from the callback
and feed it as input to another parser.

A commented-out version of a pinning interface
is present; I decline to actually implement it
(and accept its extra complexity) until/unless
it is clear that it is actually a win.  But it
is included as a proof-of-concept, to show that
it fits well with the existing interface.

+++ upb/pb/varint.h
@@ -113,7 +113,7 @@ INLINE upb_decoderet upb_vdecode_max8_fast(upb_decoderet r) {
 
 /* Encoding *******************************************************************/
 
-INLINE size_t upb_value_size(uint64_t val) {
+INLINE int upb_value_size(uint64_t val) {
 #ifdef __GNUC__
   int high_bit = 63 - __builtin_clzll(val);  // 0-based, undef if val == 0.
 #else
